+++ sdk/storage/azure-storage-files-datalake/src/datalake_service_client.cpp
@@ -53,11 +53,10 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
         if (item.Details.LeaseDuration.HasValue())
         {
           fileSystem.Details.LeaseDuration
-              = Models::LeaseDurationType((item.Details.LeaseDuration.GetValue().ToString()));
+              = Models::LeaseDuration((item.Details.LeaseDuration.GetValue().ToString()));
         }
-        fileSystem.Details.LeaseState = Models::LeaseStateType(item.Details.LeaseState.ToString());
-        fileSystem.Details.LeaseStatus
-            = Models::LeaseStatusType(item.Details.LeaseStatus.ToString());
+        fileSystem.Details.LeaseState = Models::LeaseState(item.Details.LeaseState.ToString());
+        fileSystem.Details.LeaseStatus = Models::LeaseStatus(item.Details.LeaseStatus.ToString());
 
         fileSystems.emplace_back(std::move(fileSystem));
       }