+++ sdk/storage/azure-storage-files-datalake/CHANGELOG.md
@@ -22,8 +22,9 @@
 - Removed `DataLake` from the names of return types and option types.
 - Removed `RequestId` from the return types.
 - Changed `BodyStream` parameter of `Append` function from pointer to reference.
-- Removed `PathRenameMode`, `PathGetPropertiesAction`, `PathSetAccessControlRecursiveMode`, `FileSystemResourceType` and `FileSystemResourceType`.
+- Removed `PathRenameMode`, `PathGetPropertiesAction`, `PathSetAccessControlRecursiveMode`, `FileSystemResourceType`, `PathExpiryOptions` and `FileSystemResourceType`.
 - Removed `IsAccessTierInferred` and `AccessTierChangedOn` from `PathProperties`.
+- Renamed `LeaseDurationType` to `LeaseDuration`, `LeaseStateType` to `LeaseState` and `LeaseStatusType` to `LeaseStatus`.
 
 ## 12.0.0-beta.8 (2021-02-12)